refactor(host): collapse Zeus.Server + Zeus.Desktop into single OpenhpsdrZeus binary
One executable now serves both modes — invoked as `OpenhpsdrZeus` (service:
LAN bind, HTTPS, banner) or `OpenhpsdrZeus --desktop` (Photino shell: loopback,
no HTTPS, no banner). Photino.NET moves into OpenhpsdrZeus.csproj alongside
zeus.ico/zeus.png; the BuildSpa target survives unchanged on the surviving
csproj. AssemblyName pins the output filename across all RIDs.

Namespaces stay `Zeus.Server.*` and `Zeus.Server.Hosting.*` per Phase 2a scope
— zero using-statement churn. The `public partial class Program;` anchor used
by WebApplicationFactory<Program> in tests is preserved, now hosting an
explicit [STAThread] Main that branches on `args.Contains("--desktop")`.
ZEUS_PORT and ZEUS_PERF_TEST behaviour is unchanged.

Deferred to Phase 3 (installer collapse, task #5): release.yml's per-mode
`dotnet publish` steps and the installers/*.sh / *.iss / *.cmd scripts still
reference Zeus.Server/ and Zeus.Desktop/ output paths; the build-test workflow
(which only uses Zeus.slnx) is unaffected and remains green.

Verified:
- dotnet build Zeus.slnx → 0 warnings, 0 errors
- dotnet test Zeus.slnx → 1049 passed, 137 skipped, 0 failed (matches develop)
- dotnet run --project OpenhpsdrZeus boots as service on the configured port
- dotnet run --project OpenhpsdrZeus -- --desktop boots a Photino window on a
  loopback OS-assigned port with the full host pipeline up
