--- Zeus.Desktop/Zeus.Desktop.csproj
@@ -1,49 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <!-- WinExe sets the PE Windows-GUI subsystem so launching Zeus.Desktop.exe
-         from Explorer / Start does NOT pop a console window. Console.WriteLine
-         still works when launched from cmd / PowerShell (output piped to the
-         parent's console). On macOS / Linux this is a no-op — Mach-O / ELF
-         have no PE subsystem field. -->
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <RootNamespace>Zeus.Desktop</RootNamespace>
-    <AssemblyName>Zeus.Desktop</AssemblyName>
-    <SpaRoot>$(MSBuildThisFileDirectory)..\zeus-web</SpaRoot>
-    <SpaBuildOutput>$(MSBuildThisFileDirectory)..\Zeus.Server.Hosting\wwwroot\index.html</SpaBuildOutput>
-    <!-- Embed zeus.ico into the Windows PE resource table so Explorer, the
-         taskbar, and Inno Setup's UninstallDisplayIcon all pick it up. No-op
-         on macOS / Linux where Mach-O / ELF have no equivalent slot. -->
-    <ApplicationIcon>zeus.ico</ApplicationIcon>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Zeus.Server.Hosting\Zeus.Server.Hosting.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Photino.NET" Version="4.0.16" />
-  </ItemGroup>
-
-  <!-- Ship zeus.png + zeus.ico next to the binary so Photino's SetIconFile
-       can find them at runtime — zeus.png for Linux GTK / macOS dev runs,
-       zeus.ico for Windows. CopyToOutputDirectory keeps `dotnet run` and
-       `dotnet publish` consistent. -->
-  <ItemGroup>
-    <Content Include="zeus.png" CopyToOutputDirectory="PreserveNewest" />
-    <Content Include="zeus.ico" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
-
-  <!-- Same SPA build target as Zeus.Server: zeus-web/dist → Zeus.Server.Hosting/wwwroot.
-       The library's <Content Include="wwwroot/**"> then flows the assets into our
-       publish output, so the single-process desktop app ships the SPA in-bundle. -->
-  <Target Name="BuildSpa" BeforeTargets="Publish" Condition="!Exists('$(SpaBuildOutput)') Or '$(ForceSpaBuild)' == 'true'">
-    <Message Importance="high" Text="Building React frontend in $(SpaRoot) → $(SpaBuildOutput)" />
-    <Exec Command="npm ci" WorkingDirectory="$(SpaRoot)" />
-    <Exec Command="npm run build" WorkingDirectory="$(SpaRoot)" />
-  </Target>
-
-</Project>
