--- Zeus.Server/Zeus.Server.csproj
@@ -1,36 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <SpaRoot>$(MSBuildThisFileDirectory)..\zeus-web</SpaRoot>
-    <SpaBuildOutput>$(MSBuildThisFileDirectory)..\Zeus.Server.Hosting\wwwroot\index.html</SpaBuildOutput>
-    <!-- ASP.NET Core defaults to Server GC (one GC thread per logical core),
-         tuned for high-throughput servers handling many concurrent requests.
-         Zeus.Server is a single-operator desktop radio client with a single
-         SignalR consumer; the 10 idle GC threads on a 10-core host showed up
-         as ~3% PollGCWorker overhead + a portion of LowLevelLifoSemaphore
-         wait churn in dotnet-trace cpu-sampling. Workstation GC uses one GC
-         thread and stays in step with the foreground work — lower idle
-         wake-up cost, slightly tighter pause budget per collection (Zeus
-         allocates ~1.9 MB/s so pauses are sub-millisecond). -->
-    <ServerGarbageCollection>false</ServerGarbageCollection>
-    <ConcurrentGarbageCollection>true</ConcurrentGarbageCollection>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Zeus.Server.Hosting\Zeus.Server.Hosting.csproj" />
-  </ItemGroup>
-
-  <!-- Build the React frontend into Zeus.Server.Hosting/wwwroot so it flows to
-       every downstream consumer (Zeus.Server publish output, Zeus.Desktop, etc).
-       Owning the SPA build at the executable layer (not in the library) keeps
-       'dotnet build Zeus.Server.Hosting' from shelling out to npm. -->
-  <Target Name="BuildSpa" BeforeTargets="Publish" Condition="!Exists('$(SpaBuildOutput)') Or '$(ForceSpaBuild)' == 'true'">
-    <Message Importance="high" Text="Building React frontend in $(SpaRoot) → $(SpaBuildOutput)" />
-    <Exec Command="npm ci" WorkingDirectory="$(SpaRoot)" />
-    <Exec Command="npm run build" WorkingDirectory="$(SpaRoot)" />
-  </Target>
-
-</Project>
